+++ app/src/main/java/eu/kanade/domain/base/BasePreferences.kt
@@ -33,4 +33,6 @@ class BasePreferences(
     fun displayProfile() = preferenceStore.getString("pref_display_profile_key", "")
 
     fun hardwareBitmapThreshold() = preferenceStore.getInt("pref_hardware_bitmap_threshold", GLUtil.SAFE_TEXTURE_LIMIT)
+
+    fun alwaysDecodeLongStripWithSSIV() = preferenceStore.getBoolean("pref_always_decode_long_strip_with_ssiv", false)
 }